scrapers/requirements.txt
yfinance
pandas
psycopg2-binary
requests
beautifulsoup4
python-dotenv
lxml
